pyproject.toml
[project]
name = "trae-api"
version = "0.1.0"
description = "Make trae's builtin models openai api compatible"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]

[tool.black]
line-length = 120
